+++ src/main/java/com/acme/studenthome/controller/UserAccountSystemController/UserAccountController.java
@@ -1,12 +1,9 @@
-package com.acme.studenthome.controller.UserAccountSystem;
+package com.acme.studenthome.controller.UserAccountSystemController;
 
 import com.acme.studenthome.domain.model.UserAccountSystem.Account;
-import com.acme.studenthome.domain.model.UserAccountSystem.User;
 import com.acme.studenthome.domain.service.UserAccountSystem.AccountService;
 import com.acme.studenthome.resource.UserAccountSystem.AccountResource;
 import com.acme.studenthome.resource.UserAccountSystem.SaveAccountResource;
-import com.acme.studenthome.resource.UserAccountSystem.SaveUserResource;
-import com.acme.studenthome.resource.UserAccountSystem.UserResource;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
