+++ src/main/java/com/acme/studenthome/domain/repository/LocationsSystemRepository/RegionRepository.java
@@ -1,4 +1,4 @@
-package com.acme.studenthome.domain.repository.LocationsSystem;
+package com.acme.studenthome.domain.repository.LocationsSystemRepository;
 
 import com.acme.studenthome.domain.model.LocationsSystem.Region;
 import org.springframework.data.jpa.repository.JpaRepository;
