+++ src/main/java/com/acme/studenthome/domain/repository/UserAccountSystemRepository/AccountRepository.java
@@ -1,4 +1,4 @@
-package com.acme.studenthome.domain.repository.UserAccountSystem;
+package com.acme.studenthome.domain.repository.UserAccountSystemRepository;
 
 import com.acme.studenthome.domain.model.UserAccountSystem.Account;
 import org.springframework.data.domain.Page;
